+++ cfnl.sh
@@ -4,7 +4,7 @@
 successful=false
 delete_successful=false
 triggered_delete=false
-script_version=1.6.1
+script_version=1.7.0
 # unset stack_name
 # read -p "Enter Stack Name: " stack_name
 
@@ -18,6 +18,7 @@ helpmessage="Description:
 YAML Config File Format Example:
     stackname: stack1
     profilename: awsaccount
+    region: us-west-2
     templateurl: https://s3.amazonaws.com/bucket/webapp1.yml # Or .json
     templatelocal: /path/to/cfn/templates/webapp1.yml # Unless using URL
     parametersfilepath: $HOME/.cfnl/uswest2/client1/account1/dev/webapp1.json
@@ -44,6 +45,12 @@ Examples:
     Stack Status
     $0 -s -f $HOME/.cfnl/uswest2/client1/account1/dev/webapp1.yml
 
+    Create Stack with Debugging
+    $0 -b -f $HOME/.cfnl/uswest2/client1/account1/dev/webapp1.yml
+
+    Delete Stack with Debugging
+    $0 -d -b -f $HOME/.cfnl/uswest2/client1/account1/dev/webapp1.yml
+
 Author:
     Levon Becker
     https://github.com/LevonBecker
@@ -63,15 +70,17 @@ usagemessage="Usage: $0 [-u | -d | -s] -f ./config_file.yml
 
 Options:
     -f File Path             :  (Required) YAML Script Config File Full Path
-    -u Update Stack          :  (Action Flag) Sets Action to Update Stack
-    -d Delete Stack          :  (Action Flag) Sets Action to Delete Stack
-    -s Stack Status          :  (Action Flag) Sets Action to Get Stack Status
+    -r Region                :  Overrides yaml config, ENV:AWS_REGION and aws config.
+                                Set Order: 1. Parameter 2. Yaml Config 3. ENV:AWS_REGION 4. Default us-west-2
+    -u Action Update         :  Action Flag to Update Stack
+    -d Action Delete         :  Action Flag to Delete Stack
+    -s Action Status         :  Action Flag to Get Stack Status
     -b Debug Output          :  Display Additional Output for Debugging
     -h Help                  :  Displays Help Information
     -v Version               :  Displays Script Version
 
 Action Flags:
-    Only one action flag can be used. The default Action is 'Create'.
+    Only one action flag can be used. The default Action is 'Create' (No flag).
     The three override Action Flags are -u, -d and -s.
 "
     version_message
@@ -79,11 +88,12 @@ Action Flags:
     echo "$usagemessage";
 }
 
-while getopts "f:bdhsuv" opts; do
+while getopts "f:r:bdhsuv" opts; do
     case $opts in
         b ) debug=true;;
         d ) delete=true;;
         f ) config_file_path=$OPTARG;;
+        r ) opt_region=$OPTARG;;
         h ) help_message; exit 0;;
         s ) status=true;;
         u ) update=true;;
@@ -111,9 +121,9 @@ if [ "$action_flag_count" -gt 1 ]; then
 fi
 
 if [ "$config_file_path" == "" ]; then
-usage
-echo 'ERROR: A YAML Config File is Required!'
-exit 1
+    usage
+    echo 'ERROR: A YAML Config File is Required!'
+    exit 1
 fi
 
 # Set Task Type
@@ -155,12 +165,17 @@ function show_header {
         TEMPLATE=${yaml_templatelocal}
     fi
 
-	if [ "$debug" == "true" ]; then
-	    message "** Start CloudFormation Launcher v$script_version **"
+    # If just getting Status things like ACTION are not set, so need own header
+	if [ "$status" == "true" ]; then
+        message "ACTION: STATUS"
+        message "CONFIG: ($config_file_path)"
+        message "STACKNAME: ($yaml_stackname)"
+	elif [ "$debug" == "true" ]; then
         message '** PARAMETERS **'
         message "ACTION: $ACTION"
         message "STACK NAME: $yaml_stackname"
         message "PROFILE: $yaml_profilename"
+        message "REGION: $region"
         message "TEMPLATE: $TEMPLATE"
         message "PARAMETERS FILE: $yaml_parametersfilepath"
         message "CAPABILITY IAM: $yaml_capabilityiam"
@@ -173,7 +188,6 @@ function show_header {
         message "WAIT TIME (Sec): $yaml_waittime"
         message "MAX WAITS (Loops): $yaml_maxwaits"
 	else
-	    message "** Start CloudFormation Launcher v$script_version **"
         message "ACTION: $ACTION"
         message "CONFIG: $config_file_path"
 	fi
@@ -182,18 +196,52 @@ function show_header {
 function exit_check {
     if [ "$triggered_delete" == "true" ]; then
         if [[ $1 -eq 0 || $1 -eq 255 ]]; then
-            message "REPORT: Successfully $2"
+            message "INFO: Successfully $2"
         else
             message "ERROR:  Exit Code $1 for $2"
             exit $1
         fi
     else
         if [ $1 -eq 0 ]; then
-            message "REPORT: Successfully $2"
+            message "INFO: Successfully $2"
         else
             message "ERROR:  Exit Code $1 for $2"
             exit $1
         fi
+    fi
+}
+
+function create_logs_path {
+    log_path=`dirname $yaml_logfile`
+
+    if [ ! -d ${log_path} ]
+    then
+        if [ "$debug" == "true" ]; then
+            message "DEBUG: Creating Logs Path Because Missing..."
+            message "DEBUG: Logs Path ($log_path)"
+        fi
+        mkdir -p ${log_path}
+        exit_check $? "Creating Logs Folder"
+    fi
+}
+
+function set_region {
+    if [ "$debug" == "true" ]; then
+        message "DEBUG: Setting Region..."
+    fi
+
+    if [ -n "$opt_region" ]; then
+       region="$opt_region"
+    elif [ -n "$yaml_region" ]; then
+       region="$yaml_region"
+    elif [ -n "$AWS_REGION" ]; then
+       region="$AWS_REGION"
+    else
+       region="us-west-2"
+    fi
+
+    if [ "$debug" == "true" ]; then
+        message "DEBUG: Region ($region)"
     fi
 }
 
@@ -204,22 +252,22 @@ function run_stack_command {
     elif [ "$yaml_capabilitynamediam" == "true" ]; then
         capability_iam=" --capabilities CAPABILITY_NAMED_IAM"
     else
-        capability_iam=" "
+        capability_iam=""
     fi
 
     show_header
 
     if [ "$yaml_uses3template" == "true" ]; then
-        aws cloudformation ${task_type} --profile ${yaml_profilename} \
-                                        --stack-name ${yaml_stackname}${capability_iam} \
-                                        --template-url "${yaml_templateurl}"  \
-                                        --parameters file://${yaml_parametersfilepath}
+        aws_command="aws cloudformation $task_type --profile $yaml_profilename --region $region --stack-name $yaml_stackname $capability_iam --template-url "$yaml_templateurl" --parameters file://$yaml_parametersfilepath"
     else
-        aws cloudformation ${task_type} --profile ${yaml_profilename} \
-                                        --stack-name ${yaml_stackname}${capability_iam} \
-                                        --template-body file://${yaml_templatelocal}  \
-                                        --parameters file://${yaml_parametersfilepath}
+        aws_command="aws cloudformation $task_type --profile $yaml_profilename --region $region --stack-name $yaml_stackname $capability_iam --template-body file://$yaml_templatelocal --parameters file://$yaml_parametersfilepath"
     fi
+
+    if [ "$debug" == "true" ]; then
+        message "INFO: AWS Command ($aws_command)"
+    fi
+
+    $aws_command
     exit_check $? "Executed ${ACTION} Stack Command"
     echo ''
 
@@ -227,8 +275,8 @@ function run_stack_command {
 }
 
 function delete_stack_command {
-    message 'ACTION: Deleting Stack'
-    aws cloudformation delete-stack --profile ${yaml_profilename} --stack-name ${yaml_stackname}
+    message 'INFO: Deleting Stack'
+    aws cloudformation delete-stack --profile ${yaml_profilename} --region ${region}  --stack-name ${yaml_stackname}
     exit_check $? "Executed Delete Stack Command"
 
     monitor_delete_stack_status
@@ -237,9 +285,9 @@ function delete_stack_command {
 function output_create_complete {
     # If Verbose True then Output all the Create Complete Events for Debugging
     if [ "$yaml_verbose" == "true" ]; then
-        message "REPORT: CREATE COMPLETE EVENTS..."
+        message "INFO: CREATE COMPLETE EVENTS..."
         echo ''
-        create_complete=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_COMPLETE`]')
+        create_complete=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --region ${region}  --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_COMPLETE`]')
         message "$create_complete"
         echo ''
     fi
@@ -247,9 +295,9 @@ function output_create_complete {
 
 function output_create_failed {
     # Output all the Create Failed Events for Debugging
-    message "REPORT: CREATE FAILED EVENTS..."
+    message "INFO: CREATE FAILED EVENTS..."
     echo ''
-    create_failed=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_FAILED`]')
+    create_failed=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --region ${region}  --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_FAILED`]')
     message "$create_failed"
     echo ''
 }
@@ -260,26 +308,26 @@ function monitor_delete_stack_status {
     max_waits=${yaml_maxwaits}
     while :
     do
-        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
+        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --region ${region}  --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
         exit_code=$?
         exit_check ${exit_code} "Executing Status Check"
         if [ ${exit_code} -eq 255 ]; then
-            message "REPORT: Status (DOES NOT EXIST)"
+            message "STATUS: (DOES NOT EXIST)"
         else
-            message "REPORT: Status (${STATUS})"
+            message "STATUS: (${STATUS})"
         fi
 
         if [[ "$STATUS" == "DELETE_IN_PROGRESS" && ${count} -lt ${max_waits} ]]; then
-            message "REPORT: Delete not complete!"
-            message "REPORT: Attempt $count of $max_waits."
+            message "INFO: Delete not complete!"
+            message "INFO: Attempt $count of $max_waits."
             display_runtime
-            message "Polling again in $wait_time seconds..."
+            message "INFO: Polling again in $wait_time seconds..."
             echo ''
             sleep ${wait_time}
             count=$(( count + 1 ))
         elif [ ${exit_code} -eq 255 ]; then
             delete_successful=true
-            message "REPORT: Stack Deleted ($yaml_stackname)"
+            message "INFO: Stack Deleted ($yaml_stackname)"
             break
         else
             message 'ERROR: The stack delete has failed.'
@@ -289,11 +337,11 @@ function monitor_delete_stack_status {
 }
 
 function get_stack_status {
-    STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus' 2>/dev/null)
+    STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --region ${region} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
     if [ $? -eq 255 ]; then
-        echo "Statck Status: (DOES NOT EXIST)"
+        message "STATUS: (DOES NOT EXIST) OR Token Expired"
     else
-        echo "Stack Status: ($STATUS)"
+        message "STATUS: ($STATUS)"
     fi
 }
 
@@ -302,30 +350,30 @@ function monitor_stack_status {
     max_waits=${yaml_maxwaits}
     while :
     do
-        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
+        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --region ${region} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
         exit_check $? "Executing Status Check"
-        message "REPORT: Status (${STATUS})"
+        message "STATUS: ($STATUS)"
         elapsed=$(( ($(date +%s) - ${start_time}) / 60 ))
 
         if [[ "$STATUS" == "${ACTION}_IN_PROGRESS" && ${count} -lt ${max_waits} ]]; then
-            message "REPORT: $ACTION stack is not complete!"
-            message "REPORT: Attempt $count of $max_waits."
+            message "INFO: $ACTION stack is not complete!"
+            message "INFO: Attempt $count of $max_waits."
             display_runtime
-            message "REPORT: Polling again in $wait_time seconds..."
+            message "INFO: Polling again in $wait_time seconds..."
             echo ''
             sleep ${wait_time}
             count=$(( count + 1 ))
         elif [ "$STATUS" == "${ACTION}_COMPLETE" ]; then
-            message "REPORT: $ACTION Completed!"
+            message "INFO: $ACTION Completed!"
             successful=true
             break
         elif [ "$STATUS" == "${ACTION}_FAILED" ]; then
             message "ERROR:  $ACTION Failed!"
         elif [ "$STATUS" == "UPDATE_COMPLETE_CLEANUP_IN_PROGRESS" ]; then
-            message 'REPORT: Cleanup in Progress'
-            message "REPORT: Attempt $count of $max_waits."
+            message 'INFO: Cleanup in Progress'
+            message "INFO: Attempt $count of $max_waits."
             display_runtime
-            message "REPORT: Polling again in $wait_time seconds..."
+            message "INFO: Polling again in $wait_time seconds..."
             echo ''
             sleep ${wait_time}
             count=$(( count + 1 ))
@@ -343,8 +391,8 @@ function monitor_stack_status {
                 output_create_complete
                 output_create_failed
                 message 'ERROR:  Failed and Rolling Back!'
-                message "REPORT: Rollback not complete!"
-                message "REPORT: Attempt $count of $max_waits."
+                message "INFO: Rollback not complete!"
+                message "INFO: Attempt $count of $max_waits."
                 display_runtime
                 message "Polling again in $wait_time seconds..."
                 echo ''
@@ -355,7 +403,7 @@ function monitor_stack_status {
             monitor_delete_stack_status
             break
         elif [ "$STATUS" == "ROLLBACK_COMPLETE" ]; then
-            message "REPORT: Rollback complete!"
+            message "INFO: Rollback complete!"
             echo ''
             break
         else
@@ -372,14 +420,30 @@ function display_runtime {
     message "RUNTIME: ${formatted_runtime}"
 }
 
+# Start Line here so if debugging it's before running any functions.
+message "** Start CloudFormation Launcher v$script_version **"
+if [ "$debug" == "true" ]; then
+    message "-- Debugging Enabled --"
+fi
+
 # Start Time
 start_time=$(date +%s)
+
 # Read Yaml Properties File
 eval $(parse_yaml ${config_file_path} "yaml_")
-#set | grep yaml_
+
+# Create Logs Folder if Logging
+if [ ! "$yaml_nolog" == "true" ]; then
+    create_logs_path
+fi
+
+# Set AWS Region
+set_region
+
+# Run Loop
 count=1
 if [ "$status" == "true" ]; then
-    echo "$config_file_path"
+    show_header
     get_stack_status
 elif [ "$delete" == "true" ]; then
     ACTION=DELETE
@@ -396,6 +460,7 @@ else
     run_stack_command
 fi
 
+# Footer
 if [ "$status" == "true" ]; then
     echo ''
 else
@@ -408,11 +473,11 @@ else
 
     # Results
     if [[ "$delete_successful" == "true" && "$ACTION" == "DELETE" ]]; then
-        message "REPORT: DELETE SUCCESS!"
+        message "RESULTS: DELETE SUCCESS!"
         echo ''
         exit 0
     elif [ "$successful" == "true" ]; then
-        message "REPORT: $ACTION SUCCESS!"
+        message "RESULTS: $ACTION SUCCESS!"
         echo ''
         exit 0
     else